Run the Starter process in the Docker container
Add Starter to the supervisord.conf file so that we can register a
reference element with it inside the container.  Note that the hostname
changes every time the container is run, so has to be looked up.
Also link the TANGO admin tools to /usr/local/bin so they are on the
path.

Reference:
https://github.com/3Controls/tangobox/tree/RC10/docker/base/resources

JIRA: LMC-426, LMC-443