Fix pytest collection and segmentation fault errors
Added missing files for some skabase packages to fix "collection"
errors.

Run test with pytest-xdist --boxed option, so that each in its own
process to avoid the segmentation fault.

The skabase/test/base.py module it not really used.  Moved Database()
initialisation, but not a great solution.   This file will probably
fall away in future, when tests are reworked so not a major concern.

JIRA: LMC-420