Add POGO and other Java tools
We need Pogo to do the Ansible generate-sw role.

Note that the graphical displays of these applications will not be
visible as they are handled by the offscreen Xvfb X server.  You need
to set up X11 forwarding to your host if you want to see the windows,
however this is not required for the code generation step.

The RUN step has an additional `apt-get update` to make building this
easier without having to invalidate the whole image cache.

JIRA: LMC-441