Add trivial RefCapabilityA, fix Switch simulator model key
Registering "register-my-refelt local" fails as it tries to start the
RefCapabilityA which doesn't exist, and because the simulator model
key for SvrSimSwitch is incorrect (seems the tango device names are
case sensitive).  Updated the config to fix this.

RefCapabilities B, C and D have not been implemented yet.

JIRA: LMC-441