Update license to BSD 3-clause
In preparation for transfer of this repo to the SKA Telescope
Organisation, the license is changed to BSD-3-clause.  For
simplicity, only have a single COPYING file, and not a copyright
header in every single file.

The XMI files say license "none" because POGO doesn't have a "BSD"
option.  Unfortunately when regenerating source with POGO we will
get lines like this, which aren't very helpful:
```
Distributed under the terms of the none license.
See LICENSE.txt for more info.
```

Considered using the MIT license for XMI files, but the POGO output
attributes the MIT license to someone else.