Pin pip and update setuptools in Dockerfile
Changes to pip and/or setuptools caused the building of the Docker
image to fail with an error like this:

```
error in pytango setup command: 'install_requires' must be a string
or list of strings containing valid project/version requirement
specifiers
```

Older pip and new setuptools is working.  Also removed `tango-admin`
package - not sure why it was there since it is related to Django,
not TANGO Controls.  It was also failing to install.