Fix SKABaseDevice._parse_argin test
The test with `required` was failing when calling `_init_logging` -
there is no need to call it so it was removed.

Thereafter the test passed, but incorrectly - the required parameter
needs to be a list or tuple of keys - not just a string.

Removed the SKABaseDevice.logger mock as it isn't required either.