SAR-55 Sync tango and Python logger level
This fixes the behaviour of the decorators like `@DebugIt`.
The decorators check if logging is enabled for the level of
interest by querying the Tango logger, so we need to keep
them in sync.

Note that if the Tango logger's level is set directly, (using
`set_logging_level`) then the Python logger will not be updated.
The sync is one-way: only modifying the device's `loggingLevel`
attribute sets both.