SAR-76: Use IntEnum classes for DevEnums
It is more Pythonic to use `enum.IntEnum` classes to back the
`DevEnum` attributes in Tango devices.  The base classes have now
been updated to include these.  They are all available from the
`skabase.control_model` namespace.

Some of the enumeration labels were incorrectly defined, according
to the Control System Guidelines.  These have been updated.  Also
changed labels from dashes to underscores, as that is more useful
as Python variables.  This was agreed to by Marco Bartolini.

The `TangoLoggingLevel` enum was renamed to `LoggingLevel` to match
the attribute it is linked to.  A reference with the old name is
kept for backwards compatibility.