Use python3 explicitly instead of python
The python-build docker image now has `python` pointing to Python 2.7
instead of Python 3.7, so we need to be explicit about which executable
to use.  (In the python-runtime container `python` still points to
`python3`).

Also changed pip installations to use `python3 -m pip` instead
of relying on the `pip` and `pip3` wrapper scripts.  The new approach
is recommended by pip.

Removed installation of old utility scripts from setup.py as they
are only for Python 2.7.