Add `validate_input()` method to ActionCommand.
The `validate_input()` method is called before starting the command
and calling the `do()` method, if the command has an input argument.
Overriding the new validation method is optional. Commands that fail
validation return `ResultCode.REJECTED` to the client, and no state
change occurs.  This may be preferable to returning an exception
and putting the device into a fault state.

Work initiated as part of JIRA issue: CT-166