SAR-231: Add DebugDevice command for debugpy
Allow remote debugging to be enabled by sending
a new command:  `DebugDevice`.  This will allow
attaching with a remote debugger like Visual Studio
Code.  This may degrade performance.

Currently the port cannot be changed, which would
limit usage to a single device server per Docker container.
This would be simple to change.

It has not been tested with multiple devices in a single
device server.