[SAR-227] Fix intermittent DebugDevice test failures
The debugger-related tests were failing sometimes
with a `[Errno 98] Address already in use`.
This was because debugpy was configured to listen on
the same port, 5678, across multiple tests.  Sometimes
the OS wouldn't have released the port in time for
the next test.

Now use the default port for a single test, and
ephemeral ports for the rest.