[WOM-266] Revert LRC attribute size to 64 queued commands
If a user increases the queue size of the TaskExecutorComponentManager
from the default value of 10, then we will potentially not display
information about all the queued commands unless they also increase the
maximum sizes of these attributes.

This is a pain for users to have to do so lets revert to the size of
attributes to the size we had before.