WOM-305: Create new list when updating _command_in_progress
In some situations the call to push_change_event/push_archive_event can
put this list in the omni_queue, waiting for the tango operation to be
executed later.

If a subsequent call to _update_command_in_progress modifies the list
before the operation has been executed from the omni_queue then the
original event is not sent.

This causes issues when we call Abort, where we expect an event with
('', 'Abort') to be pushed before the ('', '') event.  Without this
change, there is a possibility to receive to ('', '') events instead.