WOM-319: Ensure Abort waits for AssignResources to finish
The FakeSubarrayComponent.assign method, which was executed by the
TaskExecutor, spawns a thread to do the actual work then returns
immediately.

This means, when we come to Abort, as far as the TaskExecutor is
concerned, we are not running anything so it says that the Abort has
completed immediately.  In fact, the other thread is still running with
the actual task.

This leads to situations where the Abort command claims to be COMPLETED
before AssignResources is ABORTED.

A similar thing needs to be done for the other FakeSubarrayComponent
methods which spawn threads.