WOM-302: Fix cmd in progress push events and wait_until_done decorator
- longRunningCommandInProgress push events were referencing the same
list object in memory, which could change many times before the tango
omni thread pushes the event. Passing a new copy of the list fixes the
issue.
- In a component manager test using the FakeBaseComponent as parent
class, cmd progress events were not pushed because their on/off/standby
commands are now decorated with wait_until_done. Updated the decorator
to not use the threading event if the task_callback is a mock callable.