WOM-343: Update SKABaseDevice exception callback behaviour
The exception callback no longer updates the command result directly. If
a task callback is called with only an exception kwarg, i.e, no result
was given, a default result (ResultCode.FAILED, str(exception)) is set
in the CommandTracker and its callback is called. This prevents two
change events after each other with the same info for a task callback
provided with both a result and exception.

The exception callback now logs an Error with additional traceback and
stack info to assist with debugging.