WOM-423: Add abort method to BaseComponentManager
This abort method is used to override the behaviour of the Abort
command.  This needs to be separate from abort_tasks as abort_tasks is
the function to override if you want to change the behaviour of how
tasks are aborted in your component manager.

Also, we remove check_communicating from abort_tasks as this is not a
command implementation.  If someone overriding abort_tasks needs to
check that the device is communicating before attempting to abort the
tasks then they can add the implementation themselves, however,
implementations such as the TaskExecutorComponentManager do not need
this.

Similarly, the default abort implementation does not use
check_communicating because it only calls abort_tasks and can inherit
the check_communication status from this.