WOM-423: Revert max_executing_tasks to 2
Having `max_executing_tasks` remain as 1 when we introduce the Abort()
command changes the meaning of `max_executing_tasks`.

Before, 2739791 (WOM-344: Add Abort LRC to SKABaseDevice and let
SKASubarray override it, 2024-08-12) the `SKASubarray` device reported its
`max_executing_tasks` as 2, whereas, now it reports it as 1 because the
Abort() command is counted implicitly.

This breaks users who were overriding the `max_executing_tasks` on a
device which inherits from `SKASubarray`.  E.g. if they were overriding
`max_executing_tasks` to 3 because they can have an additional command
running in parallel, they would now have to override it to 2.

We cannot change the semantics of `max_executing_tasks` in the 1.2
release because this is a breaking change.

This reverts commit 33d3aee (WOM-344: Revert max_executing_tasks to 1,
2024-08-22).