WOM-457: Refactor LRC protocol as abstract class and class per version
The public invoke_lrc() is a factory method that finds the latest
compatible client-server protocol, instantiates a class of that
version, and calls its invoke_lrc() method. The different protocol
classes inherit from the _LrcProtocol abstract base class with common
parts that should not affect the public interface if changed. For a new
protocol version, just the invoke_lrc() and _wrap_lrc_callback() methods
need to be implemented.