WOM-501 LRCProtocol: Initialise _command_id to empty str
This avoids the following errors being printed:

push_event generated the following python exception:
Traceback (most recent call last):
  File "/home/tom/osl/skao/ska-tango-base.git/topic/wom-501-check-lrc-status/.venv/lib64/python3.10/site-packages/tango/green.py", line 249, in greener
    return executor.submit(fn, *args, **kwargs)
  File "/home/tom/osl/skao/ska-tango-base.git/topic/wom-501-check-lrc-status/.venv/lib64/python3.10/site-packages/tango/green.py", line 107, in submit
    return fn(*args, **kwargs)
  File "/home/tom/osl/skao/ska-tango-base.git/topic/wom-501-check-lrc-status/src/ska_tango_base/long_running_commands_api.py", line 391, in _wrap_lrc_callback
    cmd_idx = event.attr_value.value.index(self._command_id)
AttributeError: '_LrcProtocolV2' object has no attribute '_command_id'. Did you mean: '_command'?

This error is caused by the initial read of the attribute when first
subscribing.  This happens on the same thread as the subscription, so we
have to process it straight away and cannot wait for a command id.

This event cannot contain any information about our command as we have
not called it yet.