SKB-742: Add SlowDeviceInitCommand
The `_callback` logic in DeviceInitCommand is broken, as it results in
the "init_completed" action happening when `self._invoked()` is called,
i.e. _before_ the user supplied `do()` method.  The intended design for
this is that users call `self._completed()` themselves in their `do()`
method and this performs the "init_completed" action.

Unfortunately, fixing this would be a breaking change as users might be
relying on "init_completed" being called for them and not calling
`self._completed()` themselves.

To provide a backwards compatible fix, we introduce the
SlowDeviceInitCommand which has the correct logic.  Teams can then
inherit from this to opt-in to the correct behaviour if they need to
control when the "init_completed" action happens.