WOM-668: Monkey patch pytango attribute patching to include docstring
Without this monkey patch, if you use the named methods to provide the
read method for an attribute, then you don't get the docstring of the
method used as the description for the attribute.

This makes the named methods less convenient that using the dectorators,
however, for mix-in classes we want to prefer the name methods as they
are more easily overridden.