WOM-710: Fix long running command attribute deprecation warnings
self._version_info is overridden by users and so is not the correct
version.  It is more useful to specify the version where the deprecation
started anyway.

Also, we don't want to promise to remove them in version 2.0.0 as this
is a break in the LRC protocol and stops teams from being able to update
independently.